support defaults to be updated and local mappings to remain, small refactor
